+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file. The format 
 
 <hr />
 
+## [2.1.40] - 2025-12-25
+Summary: Update.
+
+### Changed
+- Updated staged files.
+
+### Files
+- assets/js/analyser.js
+- includes/admin/page-analyser.php
+- includes/admin/page-tools.php
+
+
 ## [2.1.39] - 2025-12-25
 Summary: Schema admin.